Minor tweaks to enums
This returns the standard stringification of enums to that of enum.Enum (i.e., a
Foo.BAR will just be "BAR"), adds qualified_name() to explicitly ask for the
full one (e.g. "Foo::BAR"), and adds a lookup() class method that takes care of
canonicalizing the input string so the lookup input can be quite flexible.