Simplification and robustness fixes to connection establishment and event retrieval
The testsuite highlighted cases where connecting the client to a newly revived
controller could hang. The unconditional get() of a status update after peering
with the controller is dangerous, since no such update may ever arrive. The
overloaded version of get() that provides a timeout currently triggers an error
in the Broker bindings (zeek/broker#241), so for now we build our own timeout
mechanism and use the status subscriber's available() method to check for
updates. This also removes the toplevel looping around the peer_nosync() call,
since it's not required: the repeated checks for status updates achieve the
same.

A second change concerns receiving events: we don't want to treat every status
update as a failure to retrieve an event. We now swallow status updates but
return actual status subscriber errors as failures to retrieve an event.

Includes unit test updates to reflect new tuning knobs.