Remove manually implemented testing framework
Python's `unittest` can discover our properly named test cases just
fine, so there is no need to implement an additional framework to
execute them.

This patch removes this useless scaffolding. In CI we run tests with
pytest to avoid the dependency on `make` and for the generally nicer
reporting, but avoid that dependency in the default setup so users do
not need to install additional development dependencies.